fix: the return type of the samples should be the same as the input (#1992)
based on @PhilipVinc suggestion

---------

+++ netket/sampler/rules/hamiltonian.py
@@ -190,7 +190,7 @@ class HamiltonianRuleJax(HamiltonianRuleBase):
         nonzero_i_plus1 = (jnp.cumsum(nonzeros, axis=-1)) * nonzeros
         rand_i_mask = nonzero_i_plus1 == jnp.expand_dims(rand_i + 1, -1)
         # .sum promotes the dtype, so we must convert it to xp dtype
-        x_proposed = (xp * jnp.expand_dims(rand_i_mask, -1)).sum(axis=1).astype(x.dtype)
+        x_proposed = (xp * jnp.expand_dims(rand_i_mask, -1)).sum(axis=1, dtype=xp.dtype)
         n_conn_proposed = self.operator.n_conn(x_proposed)
 
         # _, mels_proposed = self.operator.get_conn_padded(x_proposed)
@@ -210,7 +210,7 @@ class HamiltonianRuleJax(HamiltonianRuleBase):
 
         log_prob_corr = jnp.log(n_conn) - jnp.log(n_conn_proposed)
 
-        return x_proposed, log_prob_corr
+        return x_proposed.astype(x.dtype), log_prob_corr
 
 
 def HamiltonianRule(operator):
